Remove deprecated runtime-telemetry-java8 and runtime-telemetry-java17 modules
Functionality is provided by the unified runtime-telemetry module. Moved JarAnalyzerInstallerTest to runtime-telemetry/javaagent; JMX/JFR metric coverage is already provided by SpringBootSmokeTest and the library-level RuntimeTelemetryTest.
