+++ .fossa.yml
@@ -906,12 +906,6 @@ targets:
       target: ':instrumentation:rocketmq:rocketmq-client-5.0:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:runtime-telemetry:runtime-telemetry-java17:javaagent'
-    - type: gradle
-      path: ./
-      target: ':instrumentation:runtime-telemetry:runtime-telemetry-java8:javaagent'
-    - type: gradle
-      path: ./
       target: ':instrumentation:rxjava:rxjava-1.0:library'
     - type: gradle
       path: ./
