--- instrumentation/runtime-telemetry/runtime-telemetry-java8/javaagent/build.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-  id("otel.javaagent-instrumentation")
-}
-
-// This module's main code has been consolidated into :instrumentation:runtime-telemetry:javaagent
-// Tests are kept to verify backward compatibility with legacy property names
-
-dependencies {
-  // used to verify jar analyzer package events
-  testImplementation("commons-io:commons-io")
-  testInstrumentation(project(":instrumentation:runtime-telemetry:javaagent"))
-  testInstrumentation(project(":instrumentation:runtime-telemetry:runtime-telemetry-java17:javaagent"))
-}
-
-tasks {
-  test {
-    jvmArgs("-Dotel.instrumentation.runtime-telemetry.package-emitter.enabled=true")
-  }
-}
