--- instrumentation/runtime-telemetry/runtime-telemetry-java8/metadata.yaml
@@ -1,7 +1,0 @@
-display_name: JVM Runtime Telemetry (JMX)
-description: >
-  DEPRECATED: This instrumentation enables JVM runtime metrics using JMX (Java 8+). Use the
-  unified runtime-telemetry module instead, which provides both JMX and JFR support.
-semantic_conventions:
-  - JVM_RUNTIME_METRICS
-library_link: https://github.com/open-telemetry/opentelemetry-java-instrumentation/tree/main/instrumentation/runtime-telemetry/runtime-telemetry-java8
