+++ settings.gradle.kts
@@ -587,8 +587,6 @@ include(":instrumentation:rocketmq:rocketmq-client-5.0:javaagent")
 include(":instrumentation:runtime-telemetry:javaagent")
 include(":instrumentation:runtime-telemetry:library")
 include(":instrumentation:runtime-telemetry:testing")
-include(":instrumentation:runtime-telemetry:runtime-telemetry-java8:javaagent")
-include(":instrumentation:runtime-telemetry:runtime-telemetry-java17:javaagent")
 include(":instrumentation:rxjava:rxjava-1.0:library")
 include(":instrumentation:rxjava:rxjava-2.0:javaagent")
 include(":instrumentation:rxjava:rxjava-2.0:library")
